* src/business/business-gnome/dialog-invoice.[ch]: pass a double 	  instead of an int, because guile-1.6 sucks.  Also fix the Query 	  to actually include the Due Date of the invoice. 	* src/business/business-gnome/gw-business-gnome-spec.scm: 	  change the int to a double on the bills-due api
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@8342 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/business/business-gnome/dialog-invoice.h
@@ -33,6 +33,6 @@ GNCSearchWindow * gnc_invoice_search_edit (gpointer start, gpointer book);
 
 void gnc_business_call_owner_report (GncOwner *owner, Account *acc);
 
-DialogQueryList *gnc_invoice_show_bills_due (GNCBook *book, int days_in_advance);
+DialogQueryList *gnc_invoice_show_bills_due (GNCBook *book, double days_in_advance);
 
 #endif /* GNC_DIALOG_INVOICE_H_ */
